+++ amazonorders/entity/order.py
@@ -54,25 +54,25 @@ class Order(Parsable):
 
         # Fields below this point are only populated if `full_details` is True
 
-        #: The Order payment method.
+        #: The Order payment method. Only populated when ``full_details`` is ``True``.
         self.payment_method: Optional[str] = self._parse_payment_method() if self.full_details else None
-        #: The Order payment method's last 4 digits.
+        #: The Order payment method's last 4 digits. Only populated when ``full_details`` is ``True``.
         self.payment_method_last_4: Optional[str] = self._parse_payment_method_last_4() if self.full_details else None
-        #: The Order subtotal.
+        #: The Order subtotal. Only populated when ``full_details`` is ``True``.
         self.subtotal: Optional[float] = self._parse_subtotal() if self.full_details else None
-        #: The Order shipping total.
+        #: The Order shipping total. Only populated when ``full_details`` is ``True``.
         self.shipping_total: Optional[float] = self._parse_shipping_total() if self.full_details else None
-        #: The Order Subscribe & Save discount.
+        #: The Order Subscribe & Save discount. Only populated when ``full_details`` is ``True``.
         self.subscription_discount: Optional[float] = self._parse_subscription_discount() if self.full_details else None
-        #: The Order total before tax.
+        #: The Order total before tax. Only populated when ``full_details`` is ``True``.
         self.total_before_tax: Optional[float] = self._parse_total_before_tax() if self.full_details else None
-        #: The Order estimated tax.
+        #: The Order estimated tax. Only populated when ``full_details`` is ``True``.
         self.estimated_tax: Optional[float] = self._parse_estimated_tax() if self.full_details else None
-        #: The Order refund total.
+        #: The Order refund total. Only populated when ``full_details`` is ``True``.
         self.refund_total: Optional[float] = self._parse_refund_total() if self.full_details else None
-        #: The Order shipped date.
+        #: The Order shipped date. Only populated when ``full_details`` is ``True``.
         self.order_shipped_date: Optional[date] = self._parse_order_shipping_date() if self.full_details else None
-        #: The Order refund total.
+        #: The Order refund total. Only populated when ``full_details`` is ``True``.
         self.refund_completed_date: Optional[date] = self._parse_refund_completed_date() if self.full_details else None
 
     def __repr__(self) -> str: