+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased](https://github.com/alexdlaird/amazon-orders-python/compare/1.0.8...HEAD)
+### Added
+- Documentation improvements.
 
 ## [1.0.8](https://github.com/alexdlaird/pyngrok/compare/1.0.7...1.0.8) - 2024-01-30
 ### Added